+++ KNN/include/MinqueBase.h
@@ -6,26 +6,26 @@
 class MinqueBase
 {
 public:
-	void importY(Eigen::VectorXd Y);
-	void pushback_Vi(Eigen::MatrixXd vi);
-	void puskback_X(Eigen::MatrixXd X,bool intercept);
-	void pushback_W(Eigen::VectorXd W);
+	void importY(Eigen::VectorXf Y);
+	void pushback_Vi(Eigen::MatrixXf vi);
+	void puskback_X(Eigen::MatrixXf X,bool intercept);
+	void pushback_W(Eigen::VectorXf W);
 	void setThreadId(int Thread_id);
 	virtual void estimate()=0;
-	Eigen::VectorXd getfix() { return fix; };
-	Eigen::VectorXd getvcs() { return vcs; };
+	Eigen::VectorXf getfix() { return fix; };
+	Eigen::VectorXf getvcs() { return vcs; };
 //	void setLogfile(LOG *logfile);
 protected:
 	int nind = 0;
 	int nVi = 0;
 	int ncov = 0;
-	Eigen::VectorXd Y;
-	Eigen::MatrixXd X;
-	Eigen::VectorXd W;
-	Eigen::MatrixXd VW;
-	Eigen::VectorXd vcs;
-	Eigen::VectorXd fix;
-	std::vector<Eigen::MatrixXd> Vi;
+	Eigen::VectorXf Y;
+	Eigen::MatrixXf X;
+	Eigen::VectorXf W;
+	Eigen::MatrixXf VW;
+	Eigen::VectorXf vcs;
+	Eigen::VectorXf fix;
+	std::vector<Eigen::MatrixXf> Vi;
 //	LOG *logfile;
 	int ThreadId = 0;
 	bool iscancel=false;
